+++ src/runtime/utils/gTailwindColor.ts
@@ -1,6 +1,0 @@
-import colors from "tailwindcss/colors.js";
-export default function gTailwindColor(name:string, range:number) {
-    const gTailwindColor: any = colors;
-
-    return gTailwindColor[name][range]
-}